Iteration #1 complete

+++ accounting/models.py
@@ -3,4 +3,4 @@ from django.db import models
 
 class Expense(models.Model):
     name = models.CharField(max_length=255)
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=8, decimal_places=2)
